Creates Networking Manager
