+++ PetPicker/Controllers/ProfieViewController.swift
@@ -28,7 +28,7 @@ class ProfieViewController: UIViewController {
         user.role = roleSwitch.isOn ? "owner" : "adopter"
 
         let data = ["user": ["name": user.name, "description": user.description, "role": user.role]]
-        PPApi.shared.updateUser(data: data, id: user.id, completion: { (user) in
+        NetworkingManager.shared.updateUser(data: data, id: user.id, completion: { (user) in
             self.performSegue(withIdentifier: "savedProfile", sender: self)
         })
     }
